requirements.txt
Django==1.7.1
PyYAML==3.11
django-bootstrap-form==3.1
django-sslify==0.2.4
django-wysiwyg-redactor==0.4.3.2
gunicorn==19.1.1
selenium==2.44.0
